dse: add a way to disable the topological dimsort
This adds a utility tuple subclass `IgnoreDimSort` which can be used
to wrap the implicit dimensions tuple of an expression. When used,
this ensures that the only ordering requirements placed on the dimensions
are those in the implicit dims themselves.

It also modifies as_tuple to pass through tuple subclasses.